Moved thesis formatter to own module so it could be unit tested
